Update, fixes
Fixed #8
Fixed #7

Fixed a bug in the parsing of the domain option.

Added referer option, which works very much like the domains option, except it operates against the referer field of the headers when doing a match. Thus it is possible to have a global referer rule, or domain specific referer rules, etc etc.